This commit aims to remove a bug in how the maximum consecutive hours are evaluated in ConstraintMaxPeriodoConsecutivo :
Before, the evaluation of the number of consecutive hours was considered only if two consecutive ConcreteShifts were spotted: this way if only a single shift with duration greater than the maximum allowed period for the doctor was assigned to him, the violation would not have been noticed.

The patch just adds a check at the start of ConstraintMaxPeriodoConsecutivo.verifyConstraint that checks whether the duration of the ConcreteShift is greater than the maximum number of consecutive hours the candidate doctor can work.